Halve body/file size thresholds; refactor to fit
- Lint thresholds for function_body_length, type_body_length, file_length,
  closure_body_length halved (50/250/400/20 vs 100/500/800/40).
- Tests/.swiftlint.yml nested config relaxes thresholds for test files
  (120/600/1000) — splitting fixture-heavy tests sacrifices readability
  for marginal lint win.
- Refactor:
  - Report.swift (964L) split into Report.swift / Report+Suite.swift /
    Report+Test.swift / Report+Collections.swift.
  - Report+Convenience.swift (504L) split into Report+Convenience.swift /
    Report+Convenience+Suites.swift.
  - RepeatableTest.mergedTests (84L) broken into instance helpers
    (groupTests, mergeGroup, mergedName, mergedStatus, failureMessage,
    skipMessage, pathKey, filter).
  - SonarFormatter.format (92L) split into collectFiles / resolvePath /
    encode.
  - Report+Convenience.init (62L) extracts parseIssueMaps,
    suitesByCanonicalModule. buildFiles' nested closures lifted to a
     struct with seed/attach/build methods. processTestNode
    extracts makePathNode / pathNodeStatus.
- Two test files split into base + ExtraTests sibling to fit
  type_body_length: RepeatableTestMergedTestsTests, TestResultsDTOExtractPathsTests.
