+++ .swiftformat
@@ -53,6 +53,8 @@
 --self-required
 --init-coder-nil false
 
+--property-types inferred
+
 --decimalgrouping 3,6
 --binarygrouping 4,8
 --octalgrouping 4,8
